Worked on Profile Page and Review AJAX
The profile page now displays the user’s first and last name, a count
of likes and dislikes of beers they have made, and what beers they have
liked and disliked.  A placeholder for the reviews that the user has
made for beers is on the page.  In addition, the AJAX code has been
started for creating, editing, and deleting reviews.